refactor(prime-agent): close session and stream review gaps
