+++ packages/cli/src/daemon/openclaw.ts
@@ -8,7 +8,6 @@
 // `pendingOpenClawUiAttachJobs` is module-private working memory
 // and is intentionally not exported.
 
-import type { IncomingMessage, ServerResponse } from 'node:http';
 import { join, resolve } from 'node:path';
 import { readFile, writeFile, mkdir } from 'node:fs/promises';
 import { existsSync, readFileSync } from 'node:fs';
@@ -48,6 +47,22 @@ import {
   getStoredLocalAgentIntegrations,
   getLocalAgentIntegration,
 } from './local-agents.js';
+
+export {
+  pipeLocalAgentSseStream,
+  pipeOpenClawStream,
+  writeLocalAgentSseChunk,
+  writeOpenClawStreamChunk,
+  type LocalAgentSseReader,
+  type LocalAgentSseRequest,
+  type LocalAgentSseResponse,
+  type LocalAgentSseFrame,
+  type LocalAgentSseFrameAction,
+  type LocalAgentSseStreamResult,
+  type OpenClawStreamReader,
+  type OpenClawStreamRequest,
+  type OpenClawStreamResponse,
+} from './local-agent-sse.js';
 
 const daemonRequire = createRequire(import.meta.url);
 const execFileAsync = promisify(execFile);
@@ -562,286 +577,6 @@ export async function ensureOpenClawBridgeAvailable(
       offline: true,
     };
   }
-}
-
-export type OpenClawStreamRequest = Pick<IncomingMessage, "on">;
-export type OpenClawStreamResponse = Pick<
-  ServerResponse,
-  "on" | "off" | "writeHead" | "write" | "end" | "writableEnded"
->;
-export type OpenClawStreamReader = {
-  read: () => Promise<{ done: boolean; value?: Uint8Array }>;
-  cancel: () => Promise<unknown>;
-  releaseLock: () => void;
-};
-
-export async function writeOpenClawStreamChunk(
-  res: OpenClawStreamResponse,
-  chunk: Uint8Array,
-): Promise<void> {
-  if (res.write(chunk)) return;
-  await new Promise<void>((resolve, reject) => {
-    const cleanup = () => {
-      res.off("drain", onDrain);
-      res.off("close", onClose);
-      res.off("error", onError);
-    };
-    const onDrain = () => {
-      cleanup();
-      resolve();
-    };
-    const onClose = () => {
-      cleanup();
-      resolve();
-    };
-    const onError = (err: unknown) => {
-      cleanup();
-      reject(err);
-    };
-    res.on("drain", onDrain);
-    res.on("close", onClose);
-    res.on("error", onError);
-  });
-}
-
-/** Largest complete SSE event retained while looking for its separator. */
-const SSE_FRAME_SCAN_LIMIT = 1_000_000;
-
-const LF = 0x0a;
-const CR = 0x0d;
-
-/**
- * Byte-space match of the frame-separator grammar `/\r?\n\r?\n/` at `index`.
- * Returns the exclusive end offset of the separator, or -1 when no COMPLETE
- * separator starts here — including when a potential separator is cut off by
- * the end of `bytes`; those bytes must stay buffered until the next chunk
- * completes or disproves them, never match early.
- */
-function matchSseSeparatorAt(bytes: Uint8Array, index: number): number {
-  let i = index;
-  if (bytes[i] === CR) {
-    // Out-of-range reads yield `undefined`, so a separator truncated at the
-    // buffer end falls through to -1 without explicit length checks.
-    if (bytes[i + 1] !== LF) return -1;
-    i += 2;
-  } else if (bytes[i] === LF) {
-    i += 1;
-  } else {
-    return -1;
-  }
-  if (bytes[i] === LF) return i + 1;
-  if (bytes[i] === CR && bytes[i + 1] === LF) return i + 2;
-  return -1;
-}
-
-function findSseSeparator(
-  bytes: Uint8Array,
-  startIndex = 0,
-): { start: number; end: number } | null {
-  for (let index = startIndex; index < bytes.length; index += 1) {
-    const end = matchSseSeparatorAt(bytes, index);
-    if (end >= 0) return { start: index, end };
-  }
-  return null;
-}
-
-function concatBytes(left: Uint8Array, right: Uint8Array): Uint8Array {
-  if (left.length === 0) return right.slice();
-  if (right.length === 0) return left.slice();
-  const combined = new Uint8Array(left.length + right.length);
-  combined.set(left, 0);
-  combined.set(right, left.length);
-  return combined;
-}
-
-function parseSseJsonEvent(frame: Uint8Array): Record<string, unknown> | null {
-  const decoded = new TextDecoder().decode(frame);
-  for (const line of decoded.split(/\r?\n/)) {
-    const trimmed = line.trim();
-    if (!trimmed.startsWith("data:")) continue;
-    const data = trimmed.slice(5).trim();
-    if (!data || data === "[DONE]") continue;
-    try {
-      const parsed = JSON.parse(data);
-      return parsed && typeof parsed === "object" && !Array.isArray(parsed)
-        ? parsed as Record<string, unknown>
-        : null;
-    } catch {
-      return null;
-    }
-  }
-  return null;
-}
-
-export type LocalAgentSseFrameAction = {
-  /** Forward this frame byte-for-byte to the downstream response. */
-  forward: boolean;
-  /** Stop at this semantic frame and intentionally drop every later byte. */
-  terminal?: boolean;
-};
-
-export type LocalAgentSseFrame = {
-  event: Record<string, unknown> | null;
-  frame: Uint8Array;
-  separator: Uint8Array;
-};
-
-export type LocalAgentSseStreamResult = {
-  terminal: boolean;
-  clientGone: boolean;
-};
-
-/**
- * Canonical local-agent SSE pump.
- *
- * It owns byte-safe frame splitting, split separators, malformed/oversized
- * frames, downstream backpressure, cancellation, and terminal truncation.
- * Integrations customize only frame policy: OpenClaw/Hermes forward `final`
- * immediately, while Prime Agent withholds `error`/`final` until its durable
- * persistence hook succeeds.
- */
-export async function pipeLocalAgentSseStream(
-  req: OpenClawStreamRequest,
-  res: OpenClawStreamResponse,
-  reader: OpenClawStreamReader,
-  options: {
-    onFrame?: (
-      frame: LocalAgentSseFrame,
-    ) => LocalAgentSseFrameAction | Promise<LocalAgentSseFrameAction>;
-    endResponseOnTerminal?: boolean;
-  } = {},
-): Promise<LocalAgentSseStreamResult> {
-  let clientGone = false;
-  let terminal = false;
-  let buffer: Uint8Array = new Uint8Array(0);
-  let scanOffset = 0;
-  let oversizedFrame = false;
-  let oversizedTail: Uint8Array = new Uint8Array(0);
-  const endResponseOnTerminal = options.endResponseOnTerminal !== false;
-
-  const cancelUpstream = () => {
-    if (clientGone) return;
-    clientGone = true;
-    void reader.cancel().catch(() => {});
-  };
-
-  req.on("aborted", cancelUpstream);
-  res.on("close", () => {
-    if (!res.writableEnded) cancelUpstream();
-  });
-  res.on("error", cancelUpstream);
-
-  const handleFrame = async (
-    frame: Uint8Array,
-    separator: Uint8Array,
-  ): Promise<boolean> => {
-    const event = parseSseJsonEvent(frame);
-    const action = options.onFrame
-      ? await options.onFrame({ event, frame, separator })
-      : { forward: true, terminal: event?.type === "final" };
-    if (action.forward && !clientGone && !res.writableEnded) {
-      await writeOpenClawStreamChunk(res, concatBytes(frame, separator));
-    }
-    if (action.terminal) terminal = true;
-    return terminal;
-  };
-
-  const inspectBufferedFrames = async (atEof = false): Promise<boolean> => {
-    let boundary = findSseSeparator(buffer, scanOffset);
-    while (boundary) {
-      const frame = buffer.slice(0, boundary.start);
-      const separator = buffer.slice(boundary.start, boundary.end);
-      buffer = buffer.slice(boundary.end);
-      scanOffset = 0;
-      if (await handleFrame(frame, separator)) {
-        // The terminal frame is the stream contract. Bytes already coalesced
-        // after it belong to no subsequent turn and are deliberately dropped.
-        buffer = new Uint8Array(0);
-        scanOffset = 0;
-        return true;
-      }
-      boundary = findSseSeparator(buffer);
-    }
-
-    // The longest separator is four bytes. Everything before this point has
-    // already been disproved, while the last three bytes may be a prefix split
-    // across the next transport read.
-    scanOffset = Math.max(0, buffer.length - 3);
-
-    if (atEof && buffer.length > 0) {
-      const frame = buffer;
-      buffer = new Uint8Array(0);
-      scanOffset = 0;
-      return handleFrame(frame, new Uint8Array(0));
-    }
-
-    if (buffer.length > SSE_FRAME_SCAN_LIMIT) {
-      // A malformed giant frame must not grow memory without limit. Forward
-      // the bytes already received and scan only for its eventual separator;
-      // normal parsing resumes with the first complete frame after it.
-      if (!clientGone && !res.writableEnded) {
-        await writeOpenClawStreamChunk(res, buffer);
-      }
-      oversizedTail = buffer.slice(Math.max(0, buffer.length - 3));
-      buffer = new Uint8Array(0);
-      scanOffset = 0;
-      oversizedFrame = true;
-    }
-    return false;
-  };
-
-  const consumeOversizedFrame = async (chunk: Uint8Array): Promise<Uint8Array> => {
-    const combined = concatBytes(oversizedTail, chunk);
-    const boundary = findSseSeparator(combined);
-    if (!boundary) {
-      if (!clientGone && !res.writableEnded) await writeOpenClawStreamChunk(res, chunk);
-      oversizedTail = combined.slice(Math.max(0, combined.length - 3));
-      return new Uint8Array(0);
-    }
-    // `oversizedTail` was already forwarded. Only write the bytes contributed
-    // by this chunk through the separator, then return the unforwarded suffix
-    // to the normal frame parser.
-    const endInChunk = boundary.end - oversizedTail.length;
-    if (endInChunk > 0 && !clientGone && !res.writableEnded) {
-      await writeOpenClawStreamChunk(res, chunk.slice(0, endInChunk));
-    }
-    oversizedFrame = false;
-    oversizedTail = new Uint8Array(0);
-    return chunk.slice(Math.max(0, endInChunk));
-  };
-
-  try {
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done || clientGone) break;
-      if (value === undefined || value.length === 0) continue;
-      const unconsumed = oversizedFrame ? await consumeOversizedFrame(value) : value;
-      if (clientGone || unconsumed.length === 0) continue;
-      buffer = concatBytes(buffer, unconsumed);
-      if (await inspectBufferedFrames()) {
-        if (endResponseOnTerminal && !res.writableEnded) res.end();
-        void reader.cancel().catch(() => {});
-        break;
-      }
-    }
-    if (!clientGone && !terminal && !oversizedFrame) {
-      if (await inspectBufferedFrames(true)) {
-        if (endResponseOnTerminal && !res.writableEnded) res.end();
-        void reader.cancel().catch(() => {});
-      }
-    }
-  } finally {
-    reader.releaseLock();
-  }
-  return { terminal, clientGone };
-}
-
-export async function pipeOpenClawStream(
-  req: OpenClawStreamRequest,
-  res: OpenClawStreamResponse,
-  reader: OpenClawStreamReader,
-): Promise<void> {
-  await pipeLocalAgentSseStream(req, res, reader);
 }
 
 export function isValidOpenClawPersistTurnPayload(payload: {
