+++ packages/node-ui/src/ui/api.ts
@@ -1789,8 +1789,8 @@ interface LocalAgentChatRequestOptions {
   signal?: AbortSignal;
   identity?: string;
   sessionId?: string;
-  /** Raw framework session selected from the daemon-owned session contract. */
-  targetSessionId?: string;
+  /** One daemon-owned raw/memory pairing for an explicitly selected session. */
+  liveSession?: LocalAgentLiveSession;
   profile?: string;
   persistUserMessage?: string;
   attachments?: LocalAgentChatAttachmentRef[];
@@ -2330,14 +2330,16 @@ export interface LocalAgentIntegration {
   /** The selected Prime session currently owns an agent turn. */
   busy?: boolean;
   /** Live Prime sessions available for an explicit UI routing choice. */
-  liveSessions?: Array<{
-    /** DKG memory-session id used for history and UI selection. */
-    sessionId: string;
-    /** Prime-owned id sent to the bridge route. */
-    rawSessionId: string;
-    startedAt?: string;
-    sessionName?: string;
-  }>;
+  liveSessions?: LocalAgentLiveSession[];
+}
+
+export interface LocalAgentLiveSession {
+  /** DKG memory-session id used for history and UI selection. */
+  sessionId: string;
+  /** Prime-owned id sent to the bridge route. */
+  rawSessionId: string;
+  startedAt?: string;
+  sessionName?: string;
 }
 
 export interface LocalAgentConnectResult {
@@ -2368,7 +2370,7 @@ interface LocalAgentSurface {
   resolveChatContext?: (args: {
     integrationId: string;
     sessionId?: string;
-    targetSessionId?: string;
+    liveSession?: LocalAgentLiveSession;
     profile?: string;
   }) => Record<string, unknown>;
   fetchHealth?: () => Promise<LocalAgentHealthResponse>;
@@ -2407,7 +2409,13 @@ const LOCAL_AGENT_SURFACES: Record<string, LocalAgentSurface> = {
     // The daemon owns Prime's raw-vs-memory session contract. The UI consumes
     // its explicit fields and never reconstructs or strips the memory prefix.
     defaultSessionId: ({ record, health }) => buildPrimeAgentDefaultSessionId(record, health),
-    resolveChatContext: ({ targetSessionId }) => (targetSessionId ? { sessionId: targetSessionId } : {}),
+    resolveChatContext: ({ sessionId, liveSession }) => {
+      if (!liveSession) return {};
+      if (sessionId && liveSession.sessionId !== sessionId) {
+        throw new Error('Prime Agent live session does not match the selected conversation');
+      }
+      return { sessionId: liveSession.rawSessionId };
+    },
     fetchHealth: fetchPrimeAgentLocalHealth,
     streamChat: streamPrimeAgentLocalChat,
   },
@@ -2934,18 +2942,21 @@ export async function streamLocalAgentChat(
   const normalizedId = id.trim().toLowerCase();
   const surface = LOCAL_AGENT_SURFACES[normalizedId];
   if (surface?.streamChat) {
-    const { sessionId, targetSessionId, profile, ...transportOpts } = opts;
-    const chatOptions = (resolvedSessionId?: string, resolvedTargetSessionId?: string) => ({
+    const { sessionId, liveSession, profile, ...transportOpts } = opts;
+    const chatOptions = (
+      resolvedSessionId?: string,
+      resolvedLiveSession?: LocalAgentLiveSession,
+    ) => ({
       ...transportOpts,
       ...surface.resolveChatContext?.({
         integrationId: normalizedId,
         sessionId: resolvedSessionId,
-        targetSessionId: resolvedTargetSessionId,
+        liveSession: resolvedLiveSession,
         profile,
       }),
     });
     try {
-      return await surface.streamChat(text, chatOptions(sessionId, targetSessionId));
+      return await surface.streamChat(text, chatOptions(sessionId, liveSession));
     } catch (err) {
       // A Prime session id is an automatic live-session pin, not a durable
       // user-owned conversation id. If Prime restarted after the panel pinned
@@ -2953,7 +2964,7 @@ export async function streamLocalAgentChat(
       // once without the stale id so the daemon can elect the current head.
       if (
         normalizedId === 'prime-agent'
-        && sessionId
+        && liveSession
         && err instanceof LocalAgentApiError
         && err.code === 'PRIME_AGENT_NO_SESSION'
       ) {
