+++ packages/node-ui/src/ui/components/Shell/PanelRight/PanelRightContainer.tsx
@@ -549,17 +549,17 @@ export function PanelRight() {
       const contextEntries = [
         ...buildChatContextEntries(availableProjects, activeProjectId, currentAgent),
       ];
-      const targetSessionId = integration.id === 'prime-agent'
+      const liveSession = integration.id === 'prime-agent'
         ? integration.liveSessions?.find(
             (session) => session.sessionId === conversation.sessionId,
-          )?.rawSessionId
+          )
         : undefined;
 
       const result = await streamLocalAgentChat(integrationId, outboundText, {
         correlationId,
         signal: controller?.signal,
         sessionId: conversation.sessionId ?? undefined,
-        targetSessionId,
+        liveSession,
         profile: integration.profile,
         persistUserMessage: outboundText ? undefined : messageText,
         attachments,
